Crate partner_chains_data_source_metrics

Substrate Prometheus metrics client for Partner Chain data sources

Modules

mock

Macros

observed_async_trait

Logs each method invocation and each returned result. Has to be made at the level of trait, because otherwise #[async_trait] is expanded first. ‘&self’ matching yields “__self” identifier not found error, so “&$self:tt” is required. Works only if return type is Result.

Structs

McFollowerMetrics

Substrate Prometheus metrics client used by Partner Chain data sources

Functions

register_metrics_warn_errors

Registers new metrics with Substrate Prometheus metrics service and returns a client instance

Type Aliases

MetricsRegistry

Attribute Macros

async_trait